PHV/TPHV Privatization for read-only fields (#1430)
* Basis for PHV/TPHV privatization

This commit identifies fields that are read-only in the MAU pipeline. For such read-only fields that
are read early in the pipeline, we produce two copies of the field--one private to the MAU pipeline
(normal PHV) and the other in the TPHVs. During parsing, the input buffer contents corresponding to
the original field are written to both the normal PHV and the TPHV. The deparser emits the contents
of the TPHV copy of the field (safe since the field is never written). The MAU uses the PHV copy as
the source for instructions, match keys, etc. Since the PHV copy of the field is no longer deparsed,
we can shrink the live range for those fields, which can enable more overlay opportunities where
none existed before.

Note that:
1. This optimization requires metadata initialization to enable greater overlay.
2. The fields identified as read-only are candidates for mocha PHVs in JBay.

* Use backtracking to handle cases where privatization leads to wrong PHV allocation.

In some cases, privatized fields might cause PHV allocation to fail because of introduction of new,
unsatisfiable constraints. Such cases are handled by throwing a backtracking exception, which
invokes a second round of privatization and PHV allocation. This second round prevents the
privatization of the fields with unsatisfiable constraints.

BRIG-575 #assign #start-work #fixed