Relax constraints on deparser group threads (#1286)
* Relax constraints on deparser group gresses

The hardware requires that all deparsed containers in the same deparser group
be assigned to the same thread.  However, containers in a deparser group that
are NOT deparsed---i.e. only contain metadata---do not have to be assigned the
same thread as the rest of the group.

Prior to this commit, ALL containers in a deparser group were required to have
the same gress.  This commit relaxes that constraint to allow non-deparsed
containers to be assigned any thread.

* Add "parser group" to reflect JBay gress req.

In both Tofino and JBay, containers in a deparser group that do not contain
deparsed fields may be assigned to a thread different than the deparser group
as a whole.

However, the JBay parser views the PHV as 256 x 16b containers, which means
that even/odd pairs of 8b containers (which the parser views as one 16b
container) must be assigned to the same thread.

This commit introduces the notion of a "parser group" that is similar to
deparser groups but induced by the even/odd pairs of 8b containers.  All
containers in a parser group that contain extracted fields must be assigned to
the same thread.