Bugfix: stkvalid creation (#1497)
The trivial PHV allocator relies on PHV::Field objects having an offset
indicating their depth in the enclosing header.  POV bits have an artificial
depth calculated as if all POV bits in each gress were part of one header.

This commit fixes two problems.  First, it fixes a bug in the creation of
stkvalid fields that did not decrement the offset counter.

Second, it replaces an assert with BUG_CHECK, which guarantees that any future
bugs in calculating this offset are caught in both debug and release builds.