Ensure non-overlapping ActionAnalysis constraints (#1510)
At present, PHV allocation conservatively avoids packing two fields in the same
PHV container if they are destinations of operations that would, in turn,
require their unallocated source operands to be packed in more than one PHV
container---the analysis can determine these constraints, but the rest of
allocation can't satisfy them yet.  Instead, it assumes that all unallocated
source operands must be placed in the same container, and rejects the packing
if two or more sources are required to be in different containers.

However, it's also important to check that sources are required to be placed at
different (non-overlapping) bit positions, before merging them all into one
container placement.  This commit adds logic to check for that, and rejects the
proposed packing otherwise.