Add metadata initialization. (#1461)
This commit add a greedy algorithm that covers most of the cases of metadata initialization, while trying not introducing any new dependencies between tables.

Insert initialization as long as, it is needed, meaning that there is uninitialized read after this table,
and it is valid, in that, 1. This table will not be applied after any write to that field.
2, Field is not used in this table. If it is used in this table then initializing it here won't be able to eliminate that possible uninitialized read.

This greedy algorithm does not guarantee that all the metadata can be initialized.
There are cases that it is known not able to handle, like table that has uninitialized reads does not have any data dependency before it. In this case, we need to insert a table to initialize the field, which will extend the dependency chain. It matches what glass does right now. You can use always_init_metadata flag to make it initialize all metadata even if inserting a table dependency is required.

This algorithm will not introduce any data dependency.