packaging cleanup (#1622)
* fix condition for installing rapidjson from source

* remove all JBay references from Tofino build

* remove publishing of jbay arch files and make tofino default

* do not publish developer options

* add script to convert from P4_14 to P4_16

Adds a script to invoke the p4c-barefoot backend to pretty print
a P4_16 program. We hardcode the Tofino architecture (since we do
not expect to have P4_14 JBay programs).
We also exit early from the backend when we use the --pp argument, since
we do not want to fully compile the program, we can just run the frontend.

The script works only on the installed version of the compiler.

* increment alpha version

* add missing include

* remove more warnings

* cpplint

* add libssl dependency

* bump version to match SDE version

With the understanding that we need to tweak p4-build to accept
`--with-p4c-default`, `--with-p4c-prerelease`, and `--with-p4c-legacy` options.