Pack bridged metadata fields more compactly (#1551)
Reorder bridged metadata header in the backend

The previous bridged metadata scheme crafted a header in the midend which added padding for every
single bridged metadata field in the header. Minimizing the amount of padding requires folding in
various action-induced constraints into the creation of the bridged metadata header. While
information about uses of fields is available in the midend, other constraints such as bit-in-byte
alignment cannot be folded into the midend passes. Thus, for efficient bridged metadata packing, we
introduce a new BridgedMetadataPacking pass, which reorders the bridged metadata header in the
backend, taking into account action-induced and bit-in-byte alignment constraints.