Bit-wise aliasing + valid-->stkvalid (#1533)
* Use $stkvalid rather than $valid for stacks

Prior to this commit, we used two names as aliases for the validity bit of a
header stack element: stack[idx].$valid, but also a particular bit in
stack.$stkvalid.

This commit uses the pragma alias mechanism to hide the former and uniformly
uses the slice of the latter, which simplifies PHV allocation and header stack
manipulation.  The alias mechanism is necessary to ensure the $valid names are
reinstates in the assembly, because various other entities (eg. PD generation)
rely on those names being present.

* Allow bit-level aliasing

This commit refactors AliasMember to use an untraversed pointer to an source
IR::Expression rather than an annotation, and it adds a similar AliasSlice node
for the case when the alias destination is a slice.

It also extends the alias map to handle bit-level aliases.

Finally, it introduces helper visitors that replace AliasMembers and
AliasSlices with their sources.