Remove unnecessary extracts at the end of bridged metadata packing (#1685)
The egress::^bridged_metadata parser state in bridged metadata packing contains extracts to both the egress version of the bridged metadata fields (egress::^bridged_metadata.field_name) as well as the original egress version of the bridged field. The extracts for the former are later dead code eliminated away. This PR ensures that those extracts are removed at the end of bridged metadata packing, so that subsequent passes don't refer to unused fields, even if dead code elimination is disabled.