Support checksum in parser split. (#1699)
When create common state, checksums need to be splitted by the
amount of bytes shifted in the common state.

When split a state, if the checksum can be done within the state,
i.e. checksum reads are all within input buffer size, allocate this
checksum to this state and reserve a 16-bit extractor for it. If not,
split this checksum and set start/end/swap correctly.