Replace Field pointers in CollectBridgedFields state with field names. (#1716)
CollectPhvInfo is run multiple times since the first time CollectBridgedFields pass is run.
Therefore, the Field pointers maintained as state by this pass are stale and can (and do
occasionally) cause errors to crop up. This PR changes the field pointers to field names so that we
can get the new Field pointers (after any number of CollectPhvInfo invocations) using these names
and the PhvInfo object.