compiler output archive (#1702)
* add archive manifest

Add a compiler manifest to keep track of all compiler outputs.  This
conforms to the design in
https://docs.google.com/document/d/1yhED_blKsCnl75Sv3YbfylZ3jsGozizocGXknkzR_Ls.

Run the compiler using `--archive` to get a tar.bz2 file for all
compiler outputs. Run the compiler using `—validate-manifest` to check whether the manifest is generated according to the schema. 

Added the output directory and the program name as part of the
bf-p4c-options data structure for two reasons:
- there are several places in the code where we re-compute this information:
  - OutputAsm to generate the bfa file
  - Manifest to generate the manifest file and compute the paths
  - main p4c-barefoot function to compute graphs directory
- it also turns out that in Linux, calling dirname with a const_cast on
the 0th element of the outputFiles vector results in corrupting the options.
Therefore, we copy the cstring option into a local buffer and we invoke both
dirname and basename to get the directory and the program name.