Implement dark PHV privatization. (#1713)
The occupancy of dark containers for most programs is negligible as most fields are used in the
parser/deparser, or the input crossbar. The best utilization of dark containers requires
their use as extra PHVs available for spilling fields that are not used in particular stages.
However, this would require us to move to a stage-aware allocation. Until we make that transition,
we introduce the DarkPrivatization pass, which makes a new copy of a mocha candidate field, inserts a table in the earliest stage of each gress to initialize this copy from the original field, and
replaces all uses of the mocha field in the pipeline with this dark-privatized copy. This allows us
to allocate more fields to dark containers.
* Only enable dark privatization when JBay is specified as the target.
* Prevent insertion of initialization table if no privatized dark PHVs.
* Prevent deparsed_to_tm() fields from being privatized to dark.
* Ensure that field lists in hash dists are not allocated to dark.