Use PI branch for PI in bootstrap script
Instead of master. The PI repo is changing quite a lot right now with
the upcoming P4Runtime v1.0 release, and I cannot keep up with the
changes in bf-p4c-compilers. By introducing a stable tag, I'm giving
myself more time to update the compiler and I ensure that the new PI
updates are not too disruptive.