Add support for pa_container_type pragma. (#1727)
`@pragma pa_container_type gress field_name kind` specifies that field_name can be allocated to the
kind-type container. The pragma does not force allocation to a particular container type; rather it
only informs PHV allocation that the field `can` be allocated to that container type. This pragma
currently does not support TPHV container types.

* Add gtest.
* Force allocation of pa_container_type pragma to the specified container type.

If a field has a pa_container_type pragma attached, then raise the priority of that field's cluster
to ensure that it will be allocated to the specified container type.