Fix @name annotations on match keys (#1720)
* Fix @name annotations on match keys

Before this commit, we checked for @name annotations on match keys and used
them to set the external name for the PHV field (PHV::Field object) being
matched on.  This globally renames the field, but the @name annotation on match
keys is supposed to be local to that table.  Moreover, if the same field has
different @name annotations in different tables, only one is used.

This commit copies @name annotations from keys in the mid-end IR to annotate
InputXBar nodes in the back-end IR.  Then, the p4_param_order list in assembly
generation is drawn from these name annotations directly, if present.

To make this work, this commit also extends the assembly syntax, adding a
"key_name" property to p4_params in match tables, which, when present, is used
to produce the field names in match_key_fields and hash_bits nodes in
context.json.

* Use P4-style stacks in context.json PHV info

Assembly syntax calls for header stack elements to be referenced as
"hdr$0.field" rather than "hdr[0].field".  Ostensibly, the assembly converts
these back to P4-style names when emitting context.json.  However, it turns out
the regular expression matching in this part of the assembler was incorrect,
and, as a result, the assembly-style syntax is always emitted.

This commit (a) fixes the regex to properly match, and (b) removes these
conversions everywhere except from the PHV info, which is used for
visualization.