Support P4Info and BF-RT generation for TNA (#1748)
* Support P4Info and BF-RT generation for TNA

We support match-tables, action profiles/selectors, counters, meters and
digests.

For TNA the P4Info message generated is the Tofino-specific one, which
includes Tofino-specific extensions (for externs). Note that the P4Info
generation happens just after the frontend, before TNA translation, so
the P4Info message generated matches the original architecture used by
the P4 programmer.

The BF-RT info generation is done from the P4Info message. This means
that the generator supports both standard P4Info and Tofino-specific
P4Info. We use common C++ objects for externs which are similar across
architectures (e.g. counters, meters) to reduce code duplication. In
theory the BF-RT info generated for a PSA program and the BF-RT info
that would be generated from the same program first translated to TNA
should be the same. Whether this is true in practice is something that
needs to be verified. If the current approach shows limits, we can
always move the BF-RT info generation after the translation pass.

For the sake of testing we enable BF-RT generation for switch.p4_16.

This change required updating the p4c submodule refpoint.

BRIG-675 #fixed #fixed
BRIG-676 #fixed #fixed

* Update XFAILs & address review comments