Fix C++ iterator (was used like C-style pointer) (#1773)
C++ iterators don't work like C pointers; the end() (or rend()) method returns
an invalid iterator that points to the theoretical element preceding the first
element in the collection.  It's generally better to use
`it != collection.end()` rather than `it < collection.end()`. (And it's
incorrect to use `it <= collection.end()` and dereference `it`, because `it`
is invalid when `it == collection.end()`).