fixes infinite loop in gateway optimization (#1777)
fixes infinite loop in gateway optimization

GatewayOpt uses a PassRepeat to reach a fix point for simplifying
the gateway conditions. One of the simplifications is a normalization
of Relation expressions when one term is a constant. However, if both
are constant, the IR tree was continuously swapping the terms, thus
causing the program to change and thus never reach the fixed point.

This fix simply checks whether only the left side is constant.