Add bit-in-container alignment for SALU operands (#1755)
PHV operands to SALU instructions need to be aligned at the LSB of their PHV
containers.  Moreover, if split across multiple containers, each slice needs to
be aligned w.r.t. the positions of the other slices.  For example, a 16b field
can be split (8b/8b) across two 16b containers, as long as the lower order 8b
slice is allocated to the lower order bits of one container, and the higher
order slice is allocated to the higher order bits of another.

This commit adds a mechanism for setting valid start bits of a PHV field, and
for automatically generating the valid start bits of a slice from the valid
start bits of the field.

Eventually, this new mechanism should subsume both the bit-in-byte alignment
constraint as well as the valid container range constraint.  That will be a
future commit.