PHV Changes: Pack bridged metadata more tightly, change PHV allocation heuristic (#1785)
This PR has a bunch of changes to PHV allocation for fitting the DC_BASIC_PROFILE on switch 8.2. 
The main changes introduced in this PR are as follows:Pack bridged metadata tighter. I have improved the analysis of bridged metadata packing to ensure
that we pack bridged fields with fewer padding bits. Among new features are the ability to pack
fields with alignment constraints in the appropriate position within a byte-aligned chunk of the
bridged metadata header. I have also introduced a checkBridgedPackingConstraints() method in
ActionPhvConstraints that only packs fields together that have the same kind of source (all PHV
sources or all action data sources).
Change AllocScore heuristic. I have changed the PHV allocation heuristic to prioritize opening fewer new containers over overlay and packing bits. Because the AllocScore is the same for both field-level scores and cluster-level scores, we previously prioritized cluster allocations with more new containers, if they had more overlay or more efficient packing. The long-term solution to this is to have two types of AllocScore: one corresponding to the field granularity and the second corresponding to the cluster-level granularity.
With these changes, we now produce a binary for switch 8.2 DC_BASIC_PROFILE.