Update p4c submodule ref point
There were a few recent changes to the p4RuntimeSerializer
infrastructure that require changes to the TNA-specific control-plane
code, so even though the p4c submodule ref point was updated very
recently I decided to update it again to make sure the next person to
update it doesn't have to deal with these issues.

I also improved the gtest infrastructure in p4c so that the psa.p4
header can now be preprocessed. As a result we do not need to keep a
preprocessed version checked-in.