provides protobuf, grpc, and rapidjson installation from local server (#1845)
* provides protobuf, grpc, and rapidjson installation from local server

If you are on the intranet and can access the compiler-svr4 package
repository, you can install third party dependencies as apt packages.
Currently supported only for Ubuntu 16.04 (xenial), 32- and 64-bit.

Also provided are the scripts to generate the source packages. YMMV.

These are not tested in any of our combinations, so no need to run
Travis or Jenkins. Tested manually on a clean Ubuntu 16.04 VM.

* use netcat rather than ping to figure out whether we have access to p4c_server

Adds support for detecting linux distributions by setting three variables:
- linux_distro to identify the distribution
- linux_version is the OS release
- linux_codename is the nickname of the release.

Adds installation packages for Fedora.

Adds support for openssl > 1.1.0 when compiling grpc, and verifies
compilation using gcc-7.

Also cleanup the use of the package server for Ubuntu, as well as the
consistent use of the defined linux distribution.

Use https for connecting to the compiler server