Implement pragma pa_no_init (#1891)
* Implement pragma pa_no_init

Specifies that the indicated metadata field does not require initialization to 0 before its first
use because the control plane guarantees that, in cases where this field's value is needed, it will
always be written before it is read.  This pragma is ignored if the field’s allocation does not
require initialization. This pragma is also ignored if it is attached to any field other than a
metadata field. The gress value can be either ingress or egress. Use this pragma with care. If that
unexpected control flow path is exercised, the field will have an unknown value.

* Add Gtest

* Do gress validity checking using a static function rather than duplicating code.

BRIG-547 #bug-present #assign #assign #start-work #fixed #fixed