Do not use priority 0 for ternary matches in P4Runtime tests
According to the P4Runtime specification, 0 is not a valid priority
value for ternary match keys (while it is the only valid value for
non-ternary match keys). This is enforced by the latest version of the
P4Runtime server, therefore causing tor & fabric tests to fail. This
patch ensures that the default priority used for ternary matches is
1. It also enhances the P4RuntimeTest base PTF class to enable the test
to provide priority values when inserting ternary entries.