Missing case in pa_container_size (#1900)
* Missing case in pa_container_size

During PHV allocation, fine-grained pre-slicing first generates slices of fields based on the uses
of fields in the program (e.g. assignment using field slices, matching on field slices instead of
entire fields, etc.) However, this pre-slicing may create slices within the slice list that
individually do not satisfy the pa_container_size requirements but taken together as a slice list,
satisfy the requirements.

In the second phase of PHV allocation, we slice superclusters such that all superclusters for
allocation have widths less than or equal to the maximum container size. If any such slicing does
not satisfy the container size requirements, PHV allocation tries more slicings (slicing and
checking the validity of the resulting constraints is a very slow process) until it produces a valid
slicing or until it exceeds the number of valid slicings allowed. While checking each slicing for
unsatisfiable fields, PHV allocation checks each individual slice within a slice list, but it does
not check the slice list as a whole, thereby missing the case where a combined slice (combination of
multiple slices in the slice list that each belong to the same field) can satisfy the container size
requirements. If the combined slice list satisfies the container size requirement, then the
individual slices in the slice list do too, because all slices in the slice list are allocated to
the same container.

This PR adds support for producing and checking this combined slice list.

BRIG-759 #bug-present #assign #assign #start-work #fixed #fixed

* Add example to comments.