fix infinite loop in CanonGatewayExpr (#1904)
fix infinite loop in CanonGatewayExpr

This fixes a number of issues found by p4-smith when creating very
complex expressions. There are two fixes:
1. more aggressive simplification of expressions, for identity,
same expressions, double negation, etc. We implement these both in the
preorder visitor and the postorder to ensure that transformed expressions
are also simplified.
2. introduce a set to track history of simplified expressions so that
we do not flip-flop between different expressions that are equivalent
but can no longer be simplified.

The rest of the PR is a lot of printing for debugging with indentation,
so that we can track the expressions as they are transformed.

BRIG-765 #assign #start-work #fixed fix infinite loop in CanonGatewayExpr #fixed