Support stateful registers in BF-RT JSON
As usual this required generating the Tofino-specific information for
each Register extern in the P4Info message, then converting this
information to BF-RT JSON format.

We currently only support TNA registers and not PSA registers, since we
have not run into a use-case yet.

RegisterParam externs are not supported yet, since support is lacking in
the compiler & drivers.

BRIG-766 #fixed #fixed