Add latest version of fabric.p4 as distributed by ONF
I'm adding a new submodule ref point to track the latest version of
fabric.p4, like we do for switch.p4. We keep the old version for better
test coverage, but we can remove it once we get feature parity. The old
submodule is made "shallow", which should save ~100MB for new clones of
Brig. This only works if the bf-master-old branch of the bf-onos repo is
not updated any more, which should be the case.

The PTF tests for the latest fabric are pretty much copied from the
opennetworkinglab/fabric-p4test repository. It would not have been
convenient to add it as yet another submodule, since they are / will be
subtle differences.

The SPGW profile currently does not fit (marked as XFAIL). I will open a
JIRA once this is merged.

Hopefully at some point fabric.p4 is versionned which will make things
easier.