Add support for port metadata (aka phase0) in BF-RT JSON
With corresponding changes in the TNA P4Info message.
Milad is in the process of adding a port_metadata_extract extern method
which can be called in the parser and let the programmer specify a
custom struct for phase0 metadata. This extern method can be called at
most once. If it is not called, we still generate a table in BF-RT JSON
with a default 64-bit data field.
Once the new extern method is added to TNA, this code will need to be
adjusted slightly so that the correct programmer-provided phase0 layout
can be included in BF-RT.