Another pa_container_size fix. (#1960)
In PR-1900, I added the capability for pa_container_size to respect slicings where entire slice
lists respect the pa_container_size pragma even when the individual slices do not, by adding the
combined slice (representing the entire slice list) to the `field_slice_req_i` map that stores the
container size requirements. However, this left the case where the slices that share a rotational
cluster with the aforementioned individual slices end up failing to satisfy the pa_container_size
requirements. Due to the nature of clustering and how it interacts with the pa_container_size
slicing, we need to also add entries corresponding to the fields that are in the same rotational
cluster as the individual slices mentioned above to the `field_slice_req_i` map. This PR adds this
missing condition.

With this PR, `MSDC_IPV4_PROFILE` now passes PHV allocation. However, it fails table placement.

BRIG-802 #bug-present #assign #assign #start-work #fixed #fixed