Separate PHV analysis and allocation into distinct phases. (#1970)
* Separate PHV analysis and allocation into distinct phases.

I have now separated the analyses for PHV allocation from the transformations (cluster formation and
AllocatePHV). Previously, we set a few special constraints for mirroring-related fields and meter
color destinations at the beginning of AllocatePHV. However, this was causing the formation of
clusters that did not take these special constraints into account. This PR fixes the problem and
ensures that no new constraints are added by the time cluster formation is initiated.

* Update XFails

* Add failing test case.

This is a negative test, i.e. it is expected to fail until we relax the requirement of meter color
destinations always going to a 8-bit container.