Add support for LPF and WRED in BF-RT JSON
The code is very similar for both externs; however IMO it's a little
tricky to unify it and keep the code readable.

I added a P4_16 test programs for indirect Lpf and Wred (there was
already one for the direct versions). It fails to compile with Brig but
the BF-RT JSON is generated correctly.

BRIG-834 #fixed #fixed