Strengthen the handling of conditional constraints emitted by ActionPhvConstraints (#1981)
* Strengthen the handling of conditional constraints emitted by ActionPhvConstraints

The `copacking_constraints` UnionFind structure is generated by the `can_pack()` method in
`ActionPhvConstraints`; `copacking_constraints` consisted of multiple sets of field slices, where
each set corresponded to all source slices that must be packed in the same container. However, the
conditional constraints returned by `can_pack()` was flattened: all unallocated field slices were
forced to be a part of the same slice list, which resulted in conflicting alignment requirements,
even when slices belonged to different field slice sets in `copacking_constraints`.

The solution is to add another dimension to the conditional constraints generated by `can_pack()`.
These new conditional constraints consist of multiple maps: where each map stores the conditional
constraint information (alignment in container, container number, etc.) for all field slices that
must be allocated in a single container. By doing this, `ActionPhvConstraints` and `AllocatePHV` can
now handles allocation of superclusters where a field or a slice list is written to in multiple
actions by multiple unallocated fields/slice lists.

* Fix conditional constraint bug in earlier commit

When conditional constraints are not satisfied, then the parent transaction must abort and move to
the next container in the group.

* Avoid adding alignment requirement for smaller slice if it already has an alignment specified.

* Also, cpplint

* Resolve slice list ambiguity

The new ConditionalConstraints structure is now updated correctly when a slice is already present in
a slice list and also returned as part of conditional constraints.

* Update XFails

* Fix a bug I had introduced in ActionPhvConstraints

The alignment of the smaller slice must be ignored only when it is part of a copacking_constraints
set that has multiple slices as members.

* Update JBay timeout and XFails; address review comments

At this point, the handling of conditional constraints in ActionPhvConstraints takes most cases into
account, so I feel it's fine to close BRIG-421. There are other JIRA issues that track further
optimizations pertaining to conditional constraints.

BRIG-421 #bug-present #assign #assign #start-work #fixed #fixed