Compiler/Assembler/Infra support for 32Q  (#1967)
* Generate separate bfa for each pipe, modified driver to use bfas-runner

* update runtest script to handle 32q

* script to parse manifest

* rename and install helper scripts

Renamed scripts that get installed with a p4c- prefix, so it is clear
they are related to the p4c-compiler. This includes the manifest config
and, the bfrt conf file generator, and the bfrt schema generator.

Also added an installdir option to p4c-gen-bfrt-conf so that the paths
are adjusted for the directories where the config and context are installed.

Fixed the pipe_scope to a sensible setting, rather than all the pipes
for all contexts.