Fix corner case in ActionPhvConstraints (#1998)
For cases where we have one allocated source container, multiple unallocated sources, and some bits in the
container that are not written by an action, then the multiple unallocated sources must all be
allocated in the allocated source container. For this case, the subsequent rotational alignment
check should consider the number of sources for that action to be 1 (because we know the unallocated
and allocated sources all go in the same container).