Fix PHV allocation issues for L3_HEAVY_INT_LEAF_PROFILE (#2012)
* Record bridged fields that violate alignment constraints during PHV allocation

The INT profiles have a few cases where bridged metadata packing induces alignment constraints that
cannot be later satisfied by PHV allocation. For such cases, we can use backtracking to avoid
packing the errant fields with other bridged fields. As part of this commit, I have put in the
infrastructure to identify these fields with conflicting alignments.

* Repack bridged metdaata fields based on backtracking information

If unallocated fields involve bridged metadata fields that were not allocated because of conflicting
alignment constraints, use backtracking to communicate these field names back to the
`BridgedMetadataPacking` pass. In the subsequent rounds of `BridgedMetadataPacking`, these fields
are treated as `no-pack` with other bridged fields. The set of field names always grows with every
backtracking call, so this backtracking is guaranteed to converge (and in the worst case, every
single bridged field will never share a byte with any other bridged field).

* Update XFails. Code cleanup.