use assembler as the preprocessor language to avoid "missing terminating ' character" warning
Using C (or default) as the preprocessor langauge prints a whole
bunch of warnings when using literals in P4-14. With this change,
we tell the preprocessor to use a different language (assembler-with-cpp)
which does not interpret ' as beginning of strings.

BRIG-872 #assign #start-work #fixed use assembler as the preprocessor language to avoid "missing terminating ' character" warning #fixed