Ghost fields' PHV allocation related fixes (#2073)
The hardware writes ghost metadata to a single 32-bit phv container in a very constrained form. The
fields are written to certain bits in a fixed way:
- pipe to [1:0]
- qid to [12:2]
- qlength to [30:13]
- pingpong to [31:31]

So phv allocation must match that. If a particular field is unused (pipe is rarely used), the
hardware will still write those bits, so it may not be possible to reuse it for something else (need
to add an explicit write in the mau pipe). The fields also may not be relocated within the
container.

BRIG-899 #bug-present #assign #assign #start-work #fixed #fixed