fix manifest issues (#2112)
* fix manifest issues

With T32q we need to publish the names of the control flows.

We hacked this to override the pipe field, however, that is needed
by p4i to display the pipelines. So this PR adds a separate label
for the control flow name and updates the p4c-manifest-config script
to use it.

BRIG-910  #assign #start-work #fixed #fixed

In addition, we also fix the traversal of control blocks to generate
entries for all graphs in the manifest file.

BRIG-911  #assign #start-work #fixed #fixed

And finally, update the schema version for proper validation.

* fix bfas-runner to look for pipe_name, add enum, and always validate the manifest (in developer mode)

* integrate manifest parser into p4c driver

Override the run method of the driver to process the manifest file
and invoke the assembler multiple times.

Also fix the case for enumeration in the schema, 'cause it wouldn't
verify for p4-16. We do need to add p4-16 tests.

BRIG-880  #assign #start-work #fixed #fixed

* add compiler flags to manifest schema

And bump the schema version up, even though we are not currently
generating them.

* preserve arguments passed on the cmd line for the assembler

At the same time, generate only one pipe for JBay, as well as
--singlewrite to reduce the size of the blob we load in the driver.