add logging to files (#2172)
* add p4c project key

* add logging to files

Given that customers are expecting to check the logs when allocations
fail, it is desirable to log allocation steps to files. On a failed
compilation they can search through the logs to find issues.

The current PR is effectively a stop gap until we provide structured
logging and display the logs in p4i. For now, we simply redirect the
output of clog to a file (table_placement_x.log, phv_allocation_x.log,
table_summary_x.log). These are the logging options that we provide on
our wiki. We should consider other logs as well.