split resources node into a separate .res.json file (#2177)
split resources node into a separate .res.json file

Since we now have the manifest to manage a set of files,
we can split the resources node into its own file and remove it
from context.json. This keeps the context smaller as bf-driver
does not need to see the resources.

Additions:
- Add resource validation in addition to context.json
  validation. Refactored the code for the validator into a
  JSONValidator class.
- Add fields for resources.json: compiler_version, build_date, program_name.
- Advance the manifest schema version.
- Add `--archive` name argument. Add an optional name to the archive,
so that users can create different archives for the same program and
add a few compiler options tests to validate it.
- Make `Logging::Manifest` a singleton object. We want to call this
from anywhere in the compiler. In particular we want to use it for
adding logs.
- Add ghost thread as enum to schemas
- And last but not least, fix the output of the manifest for P4-14.